Unify function interfaces

convoke unifies functions along a single specification using statements on function argument transformations that would be needed. The result is a single function whose interface can be changed simply by changing the respective argument in the generated function.

convoke(uniface, ...)

Arguments

uniface
[function] The desired unifying interface. Function returning a string as description

...
[function ~ purrr_lambda] All the various specifications for function argument transformations.

Value

Function with additional class convoke with arguments being

convoke_func(specified_args, interface, ..., evaluate = TRUE)

The evaluate argument is useful for debugging purposes. Extra arguments can be passed in the format interface.arg.

Unifying Functions

Specifying the unifying interface

The unifying interface needs to be specified in the format

\(arg1 = default1, arg2 = default2, etc.) "description"

This uses R's recent anonymous function syntax.

Specifying function argument transformations

Transformations follow the formula format

func1(func1arg1 = transformed_arg1, func1arg2 = transformed_arg2) ~
postprocess(result)

Essentially, pass the function as if it is to take arguments from the unified function.

Postprocessing function

The postprocess function on the right of the formula specifies what postprocessing is required on the result of applying func1. It is to be specified similarly to how functions in purrr::map are specified (that is by the . and .x syntax.

Composing functions

It is also possible to progressively add functions to a convoke function simply by adding to the convoke function the new specifications:

convoke_func + (func_spec ~ postfunc)

See also

Examples

# unifying functions with swapped arguments
foo <- function(a1, b1) { a1/b1 }
bar <- function(b2, a2) { a2/b2 }
convoked <- convoke(\(a, b) "",
                   foo(a1 = a, b1 = b) ~ .,
                   bar(b2 = b, a2 = a) ~ .)
c(convoked(9,5), bar(5,9))
#> [1] 1.8 1.8
c(convoked(9,5,interface="bar"), foo(9,5))
#> [1] 1.8 1.8

# supplying optional arguments
foo <- function(a1, b1, round=FALSE) { if(round) round(a1/b1) else a1/b1 }
bar <- function(b2, a2) { a2/b2 }
convoked <- convoke(\(a, b) "",
                   foo(a1 = a, b1 = b) ~ .,
                   bar(b2 = b, a2 = a) ~ .)
c(convoked(9,5, foo.round=FALSE), bar(5,9))
#> [1] 1.8 1.8
c(convoked(9,5, foo.round=TRUE), round(bar(5,9)))
#> [1] 2 2

# post-processing specification
foo <- function(a1, b1) { a1/b1 }
bar <- function(b2, a2) { a2/b2 }
multfactor <- 2
convoked <- convoke(\(a, b) "", 
                   foo(a1 = a, b1 = b) ~ .*multfactor,
                   bar(b2 = b, a2 = a) ~ .*multfactor)
c(convoked(9,5), bar(5,9)*multfactor)
#> [1] 3.6 3.6
multfunc <- function(x) x*2
convoked <- convoke(\(a, b) "", 
                   foo(a1 = a, b1 = b) ~ multfunc(.),
                   bar(b2 = b, a2 = a) ~ multfunc(.))
c(convoked(9,5), bar(5,9)*multfactor)
#> [1] 3.6 3.6

# adding further functions later
foo <- function(a1, b1) { a1/b1 }
bar <- function(b2, a2) { a2/b2 }
baz <- function(a3, b3) { a3/b3 }
(convoked <- convoke(\(a, b) "", foo(a1 = a, b1 = b) ~ .))
#> convoke function 
#>   interfaces: foo()
#>   args: a, b, interface = foo, interface.args
(convoked <- convoked + (bar(b2 = b, a2 = a) ~ .) +
    (baz(a3 = a, b3 = b) ~ .))
#> convoke function 
#>   interfaces: foo(), bar(), baz()
#>   args: a, b, interface = foo, interface.args
c(convoked(9,5, interface="baz"), bar(5,9))
#> [1] 1.8 1.8